c++, coroutines: Make and use a frame access helper.
In the review of earlier patches it was suggested that we might make
use of finish_class_access_expr instead of doing a lookup for the
member and then a build_class_access_expr call.

finish_class_access_expr does a lot more work than we need and ends
up calling build_class_access_expr anyway.  So, instead, this patch
makes a new helper to do the lookup and build and uses that helper
everywhere except instances in the ramp function that we are going
to handle separately.

gcc/cp/ChangeLog:

	* coroutines.cc (coro_build_frame_access_expr): New.
	(transform_await_expr): Use coro_build_frame_access_expr.
	(transform_local_var_uses): Likewise.
	(build_actor_fn): Likewise.
	(build_destroy_fn): Likewise.
	(cp_coroutine_transform::build_ramp_function): Likewise.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>