arm: [MVE intrinsics] use long_type_suffix / half_type_suffix helpers
In several places we are looking for a type twice or half as large as
the type suffix: this patch introduces helper functions to avoid code
duplication. long_type_suffix is similar to the SVE counterpart, but
adds an 'expected_tclass' parameter.  half_type_suffix is similar to
it, but does not exist in SVE.

2024-08-28  Christophe Lyon  <christophe.lyon@linaro.org>

	gcc/

	* config/arm/arm-mve-builtins-shapes.cc (long_type_suffix): New.
	(half_type_suffix): New.
	(struct binary_move_narrow_def): Use new helper.
	(struct binary_move_narrow_unsigned_def): Likewise.
	(struct binary_rshift_narrow_def): Likewise.
	(struct binary_rshift_narrow_unsigned_def): Likewise.
	(struct binary_widen_def): Likewise.
	(struct binary_widen_n_def): Likewise.
	(struct binary_widen_opt_n_def): Likewise.
	(struct unary_widen_def): Likewise.