testsuite, c++, coroutines: Correct a test intent.
The intention of the series of tests numberef pr95615-* is to
verify that entities created by the ramp and potentially needing
destruction are correctly handled when exceptions are thrown.
Because of a typo, one case was not being checked correctly (the
return object).  This patch amends the check to test that the
returned object is properly deleted.

gcc/testsuite/ChangeLog:

	* g++.dg/coroutines/torture/pr95615.inc: Check tha the
	task object produced by get_return_object is correctly
	deleted on exception.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>